pyproject.toml
[project]
name = "vk-to-tg"
version = "0.1.0"
dependencies = [
    "ruff",
    "vk-api",
    "python-dotenv",
    "yt-dlp",
    "telethon",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
